requirements.txt
line-bot-sdk
flask
json_minify
requests
bs4